requirements.txt
setuptools
wheel
build
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
requests==2.31.0
openai==1.3.7
beautifulsoup4==4.12.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
fastapi
uvicorn
langchain
requests
python-dotenv
sentence-transformers
faiss-cpu